miscellaneous-scripts/export_to_s3/app/database/conn_mgr.py
import psycopg2
import os


DB_NAME = os.getenv('DB_NAME')
DB_USER = os.getenv('DB_USER')
DB_PASS = os.getenv('DB_PASS')
DB_HOST = os.getenv('DB_HOST')
DB_PORT = os.getenv('DB_PORT')


def get_db_connection():
    """get connection of annotation tracking database"""

    conn = psycopg2.connect(database=DB_NAME, user=DB_USER, password=DB_PASS,
                            host=DB_HOST, port=DB_PORT)
    return conn


def close_conn(conn):
    """close connection of database"""
    conn.close()
