miscellaneous-scripts/requirements.txt
aiofiles==0.8.0
anyio==3.6.1
argcomplete==2.0.0
backports.cached-property==1.0.2
backports.zoneinfo==0.2.1
boto3==1.24.40
botocore==1.27.40
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
dacite==1.6.0
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
eventlet==0.33.1
fiftyone==0.16.5
fiftyone-brain==0.8.2
fiftyone-db==0.3.0
fonttools==4.36.0
future==0.18.2
glob2==0.7
graphql-core==3.1.7
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imageio==2.21.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mlutils==0.2.0b0
mongoengine==0.20.0
motor==2.5.1
ndjson==0.3.1
networkx==2.8.5
numpy==1.23.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
patool==1.12
Pillow==9.2.0
plotly==4.14.3
pprintpp==0.4.0
priority==2.0.0
psutil==5.9.1
pycocotools==2.0.4
Pygments==2.13.0
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
retrying==1.3.3
rfc3986==1.5.0
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
sentinel==0.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sse-starlette==0.10.3
sseclient-py==1.7.2
starlette==0.16.0
strawberry-graphql==0.96.0
tabulate==0.8.10
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.11
pypdf2==2.11.2
xmltodict==0.13.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5